Generate an id for each new recording session
In case of a group, the group id was used as the record session id.
Consequently each record was removing the previous one as the id is used
to store information in the device's hard drive.
It now generates a new session id each time a record session is made
avoiding collisions.