client: fix drag & drop on devices
Drag & drop devices was broken after entering and
leaving a group, the dragged device was moving
away from the mouse. The bug appeared after
upgrading interactjs the library responsible for
the drag & drop.
It was due to the drag events being set several
times. Event listeners are now removed when
destroying the controller.