feature: add support for weekly schedules
It was possible to plan a daily record, it is now
possible to also plan a weekly record.

It also contains a fix to detect a collision on
daily records which wasn't detected before.
Collision can appear when the record is planned to
start on a day and terminates the next day.