client: fix record name column in schedules table
The column of the planification table didn't have
a fixed size and was unlarging the table making
the last column leaving the screen when a word
was too long in the record name.
Column has now a fixed width and the words in the
record name can be broken to fit inside the
specified width.