client: remove end date on non recurrent schedule
Clicking on the checkbox to add a daily or weekly
record then unchecking the box was settings the
"recurrent" property to "false" in the database so
the end date could be displayed in the
planification table for non recurrent schedules.
The recurrent property is now absent from the
document when not daily nor weekly.