server/client: fix end date in collision algorithm
A collision wasn't detected by the collision
algorithm with recurrent schedules. In fact the
time of the end date of recurrent schedules was
considered to be 00:00:00 instead of the last
occurence end date / time so a schedule
terminating the same date as the end date
of a recurrent one and occupying the same ranges
of time in this end date wasn't considered as in
collision.
It is now.