dependencies: remove grunt concat and uglify
Grunt uglify and grunt concat were used to build
the JavaScript files of the back office client.
UglifyJS is now called directly and also concats
files so both grunt-contrib-concat and
grunt-contrib-uglify aren't needed anymore.

Note that the command for building the project
hasn't changed and is still `npm run build`.