client: fix control panel scrollbar
Scrollbars of Action, Detail and History tabs of
the control panel weren't working when closing the
control panel and reopening it with the same
device.

Scrollbars were destroyed when closing the control
panel but they shouldn't be because the control
panel is always displayed, they just needed to be
reset.